Purpose:     Bug fix
Problem:
    In vg.h, the constant _HDF_VDATA was re-defined to be -1.

    There are several places failing in the tester which call the attribute function

    if (FAIL == VSsetattr(vsid, (int32)_HDF_VDATA, ATTNAME3, DFNT_CHAR8, 3, attr3)) {
	...
    }

Solution:
    Take off the (int32) datatype changer.  Tester now reads:

    if (FAIL == VSsetattr(vsid, _HDF_VDATA, ATTNAME3, DFNT_CHAR8, 3, attr3)) {
	...
    }


Platform tested:
    Cray T90 - old floating point format

+++ hdf/test/tvattr.c
@@ -238,7 +238,7 @@ static intn write_vattrs(void)
          num_errs++;
          printf(">>> Search for non-existing field, should fail.\n");
       }
-   if (FAIL == VSsetattr(vsid, (int32)_HDF_VDATA, ATTNAME3, DFNT_CHAR8,3, attr3)) {
+   if (FAIL == VSsetattr(vsid, _HDF_VDATA, ATTNAME3, DFNT_CHAR8, 3, attr3)) {
          num_errs++;
          printf(">>> VSsetattr3 failed\n");
       }
@@ -270,7 +270,7 @@ static intn write_vattrs(void)
       }
 
    /* set same attr name to different fields */
-   if (FAIL == VSsetattr(vsid, (int32)_HDF_VDATA, ATTNAME4, DFNT_FLOAT32,1, attr4)) {
+   if (FAIL == VSsetattr(vsid, _HDF_VDATA, ATTNAME4, DFNT_FLOAT32,1, attr4)) {
          num_errs++;
          printf(">>> VSsetattr6 failed\n");
       }
@@ -342,7 +342,7 @@ static intn write_vattrs(void)
          printf(">>> VSsetattr6 failed\n");
       }
 
-   if (FAIL == VSsetattr(vsid, (int32)_HDF_VDATA, ATTNAME7, DFNT_CHAR8,3, attr3)) {
+   if (FAIL == VSsetattr(vsid, _HDF_VDATA, ATTNAME7, DFNT_CHAR8,3, attr3)) {
          num_errs++;
          printf(">>> VSsetattr7 failed\n");
       }
@@ -610,13 +610,13 @@ static intn read_vattrs(void)
          printf(">>> Wrong num of Vsname1 attrs, should be %d, ",6);
          printf("got %d.\n", n_vsattrs);
    }
-   if (FAIL == (n_fldattrs = VSfnattrs(vsid, (int32)_HDF_VDATA)) || n_fldattrs != 2) {
+   if (FAIL == (n_fldattrs = VSfnattrs(vsid, _HDF_VDATA)) || n_fldattrs != 2) {
         num_errs++;
         printf(">>> Wrong num of Vsname1 vdata attrs, ");
         printf("should be %d, got %d.\n ", 2, n_fldattrs);
    }
    /* look for non-existing attr, should fail */
-   if (FAIL != (iattrindex = VSfindattr(vsid, (int32)_HDF_VDATA, ATTNAME9))) {
+   if (FAIL != (iattrindex = VSfindattr(vsid, _HDF_VDATA, ATTNAME9))) {
         num_errs++;
         printf(">>> attname9 is not an attr of vdata vsname1, ");
         printf(" should fail.\n");
@@ -650,13 +650,13 @@ static intn read_vattrs(void)
    }
 
    /* get the 2nd attr  */
-   if ((FAIL == (iattrindex = VSfindattr(vsid, (int32)_HDF_VDATA, ATTNAME4))) ||
+   if ((FAIL == (iattrindex = VSfindattr(vsid, _HDF_VDATA, ATTNAME4))) ||
              (iattrindex != 1))  {
         num_errs++;
         printf(">>> attname4 should be index 1 of vsname1, not %d.\n",
                      iattrindex);
    }
-   if ((FAIL == VSattrinfo(vsid, (int32)_HDF_VDATA, iattrindex, iattrname, 
+   if ((FAIL == VSattrinfo(vsid, _HDF_VDATA, iattrindex, iattrname, 
                &i_type, &i_count, &i_size)) || 
                (HDstrcmp(iattrname, ATTNAME4) != 0) ||
                (i_type != DFNT_FLOAT32) || (i_count != 1) || 
@@ -665,7 +665,7 @@ static intn read_vattrs(void)
         printf(">>> Wrong attrinfo for attname4 of vdata vsname1; ");
         printf(" got  %s %d %d.\n", iattrname, (int)i_type,(int)i_count);
    }
-   if (FAIL == VSgetattr(vsid, (int32)_HDF_VDATA, 1, iattr4) ||
+   if (FAIL == VSgetattr(vsid, _HDF_VDATA, 1, iattr4) ||
        (fabs((double)(iattr4[0] - attr4[0])) > fabs((double)(attr4[0]*EPS32))))  {
          num_errs++;
          printf(">>> Wrong values for attname4  of vsname1; \
